+++ CHANGELOG.md
@@ -10,6 +10,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
 
+## [0.2.0] - 2024-12-06
+
+[Compare with previous version](https://github.com/sparkfabrik/terraform-kubernetes-application-sleep-cycles/compare/0.1.0...0.2.0)
+
+- Add default protected Kubernetes system namespaces (kube-node-lease, kube-public, kube-system)
+- Add management of statefulsets sleep cycle
+- Add management of all resources in a namespace
+- Refactor working-hours.sh script for better maintainability 
+
 ## [0.1.0] - 2024-11-06
 
 - First release.